+++ libs/i18n/assets/en.json
@@ -164,68 +164,68 @@
   "Next": "Next",
   "Save": "Save",
   "Are you sure you want to delete?": "Are you sure you want to delete?",
-  "Beans": "Beans",
-  "Cassava": "Cassava",
-  "Cotton": "Cotton",
-  "Cowpeas": "Cowpeas",
   "Demo": "Demo",
   "Edit Profiles": "Edit Profiles",
   "Extension": "Extension",
   "Farmer": "Farmer",
-  "Groundnuts": "Groundnuts",
   "Hide Options": "Hide Options",
-  "Maize": "Maize",
   "Name": "Name",
   "New Profile": "New Profile",
-  "Pigeon Peas": "Pigeon Peas",
-  "Pumpkins": "Pumpkins",
-  "Rice": "Rice",
   "Select Country": "Select Country",
   "Select Language": "Select Language",
   "Show Options": "Show Options",
+  "Who is using?": "Who is using?",
+  "Beans": "Beans",
+  "Cassava": "Cassava",
+  "Cotton": "Cotton",
+  "Cowpeas": "Cowpeas",
+  "Groundnuts": "Groundnuts",
+  "Maize": "Maize",
+  "Pigeon Peas": "Pigeon Peas",
+  "Pumpkins": "Pumpkins",
+  "Rice": "Rice",
   "Sorghum": "Sorghum",
   "Soya Beans": "Soya Beans",
   "Sunflower": "Sunflower",
   "Sweet Potatoes": "Sweet Potatoes",
   "Tobacco": "Tobacco",
-  "Who is using?": "Who is using?",
-  "Mon": "Mon",
   "Monday": "Monday",
-  "Tue": "Tue",
   "Tuesday": "Tuesday",
-  "Wed": "Wed",
   "Wednesday": "Wednesday",
-  "Thu": "Thu",
   "Thursday": "Thursday",
-  "Fri": "Fri",
   "Friday": "Friday",
-  "Sat": "Sat",
   "Saturday": "Saturday",
-  "Sun": "Sun",
   "Sunday": "Sunday",
-  "Jan": "Jan",
+  "Mon": "Mon",
+  "Tue": "Tue",
+  "Wed": "Wed",
+  "Thu": "Thu",
+  "Fri": "Fri",
+  "Sat": "Sat",
+  "Sun": "Sun",
   "January": "January",
-  "Feb": "Feb",
   "February": "February",
-  "Mar": "Mar",
   "March": "March",
-  "Apr": "Apr",
   "April": "April",
+  "June": "June",
+  "July": "July",
+  "August": "August",
+  "September": "September",
+  "October": "October",
+  "November": "November",
+  "December": "December",
+  "Jan": "Jan",
+  "Feb": "Feb",
+  "Mar": "Mar",
+  "Apr": "Apr",
   "May": "May",
   "Jun": "Jun",
-  "June": "June",
   "Jul": "Jul",
-  "July": "July",
   "Aug": "Aug",
-  "August": "August",
   "Sep": "Sep",
-  "September": "September",
   "Oct": "Oct",
-  "October": "October",
   "Nov": "Nov",
-  "November": "November",
   "Dec": "Dec",
-  "December": "December",
   "All": "All",
   "Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.": "Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station.",
   "Chance of receiving the water requirement in the days to maturity for this crop variety": "Chance of receiving the water requirement in the days to maturity for this crop variety",
