+++ libs/i18n/src/hardcoded/index.ts
@@ -5,8 +5,14 @@ import { ITranslationEntry } from '../types';
 import { BUDGET_ENTRIES } from './budget';
 import { CLIMATE_ENTRIES } from './climate';
 import { COMMON_ENTRIES } from './common';
+import { DATA_ENTRIES } from './data';
 
-export const HARDCODED_DATA: ITranslationEntry[] = [...BUDGET_ENTRIES, ...CLIMATE_ENTRIES, ...COMMON_ENTRIES];
+export const HARDCODED_DATA: ITranslationEntry[] = [
+  ...BUDGET_ENTRIES,
+  ...CLIMATE_ENTRIES,
+  ...COMMON_ENTRIES,
+  ...DATA_ENTRIES,
+];
 
 const PROJECT_ROOT = resolve(__dirname, '../../../../');
