+++ libs/i18n/templates/_template.csv
@@ -164,68 +164,68 @@ tool,context,text
 "common","buttons","Next"
 "common","buttons","Save"
 "common",,"Are you sure you want to delete?"
-"common",,"Beans"
-"common",,"Cassava"
-"common",,"Cotton"
-"common",,"Cowpeas"
 "common",,"Demo"
 "common",,"Edit Profiles"
 "common",,"Extension"
 "common",,"Farmer"
-"common",,"Groundnuts"
 "common",,"Hide Options"
-"common",,"Maize"
 "common",,"Name"
 "common",,"New Profile"
-"common",,"Pigeon Peas"
-"common",,"Pumpkins"
-"common",,"Rice"
 "common",,"Select Country"
 "common",,"Select Language"
 "common",,"Show Options"
-"common",,"Sorghum"
-"common",,"Soya Beans"
-"common",,"Sunflower"
-"common",,"Sweet Potatoes"
-"common",,"Tobacco"
 "common",,"Who is using?"
-"common_days","day_1","Mon"
+"common_crops",,"Beans"
+"common_crops",,"Cassava"
+"common_crops",,"Cotton"
+"common_crops",,"Cowpeas"
+"common_crops",,"Groundnuts"
+"common_crops",,"Maize"
+"common_crops",,"Pigeon Peas"
+"common_crops",,"Pumpkins"
+"common_crops",,"Rice"
+"common_crops",,"Sorghum"
+"common_crops",,"Soya Beans"
+"common_crops",,"Sunflower"
+"common_crops",,"Sweet Potatoes"
+"common_crops",,"Tobacco"
 "common_days","day_1","Monday"
-"common_days","day_2","Tue"
 "common_days","day_2","Tuesday"
-"common_days","day_3","Wed"
 "common_days","day_3","Wednesday"
-"common_days","day_4","Thu"
 "common_days","day_4","Thursday"
-"common_days","day_5","Fri"
 "common_days","day_5","Friday"
-"common_days","day_6","Sat"
 "common_days","day_6","Saturday"
-"common_days","day_7","Sun"
 "common_days","day_7","Sunday"
-"common_months","month_01","Jan"
+"common_days","day_short_1","Mon"
+"common_days","day_short_2","Tue"
+"common_days","day_short_3","Wed"
+"common_days","day_short_4","Thu"
+"common_days","day_short_5","Fri"
+"common_days","day_short_6","Sat"
+"common_days","day_short_7","Sun"
 "common_months","month_01","January"
-"common_months","month_02","Feb"
 "common_months","month_02","February"
-"common_months","month_03","Mar"
 "common_months","month_03","March"
-"common_months","month_04","Apr"
 "common_months","month_04","April"
-"common_months","month_05","May"
-"common_months","month_06","Jun"
 "common_months","month_06","June"
-"common_months","month_07","Jul"
 "common_months","month_07","July"
-"common_months","month_08","Aug"
 "common_months","month_08","August"
-"common_months","month_09","Sep"
 "common_months","month_09","September"
-"common_months","month_10","Oct"
 "common_months","month_10","October"
-"common_months","month_11","Nov"
 "common_months","month_11","November"
-"common_months","month_12","Dec"
 "common_months","month_12","December"
+"common_months","month_short_01","Jan"
+"common_months","month_short_02","Feb"
+"common_months","month_short_03","Mar"
+"common_months","month_short_04","Apr"
+"common_months","month_short_05","May"
+"common_months","month_short_06","Jun"
+"common_months","month_short_07","Jul"
+"common_months","month_short_08","Aug"
+"common_months","month_short_09","Sep"
+"common_months","month_short_10","Oct"
+"common_months","month_short_11","Nov"
+"common_months","month_short_12","Dec"
 "crop-probability",,"All"
 "crop-probability",,"Calculated by using FAO CLIMWAT 2.0 for Cropwat and Cropwat 8.0 and climate data for Nkhotakota Met Station."
 "crop-probability",,"Chance of receiving the water requirement in the days to maturity for this crop variety"
